Enables safe Mongo persistence
Right now MongoDB's maximum document size is 16Mb. A large page encountered
during a crawl can raise BSON::InvalidDocument an interrupt the crawl.

This commit implements a safe storage strategy which watches for this
condition and, when encountered, clears the body, marks the document as
too large and attempts to save it again.